// Front_agriculture intelligente/Front_agriculture intelligente/src/app/model/device.ts
export  class Device {
  _id: string;
  userid: string;
  name: string;
  lng: number;
  lat: number;
  state: boolean;
  siteId: string;
  siteName: string;
  ligneName: string;
  ligneId: string;
code: string;
  type: string;
  data: [];
  Consomation: [];
  Countersdata: [];
  ConsomationTripahse: [];

  PositiveTripahse: [];

  ReverserTipahse: [];

  ActivePowerTipahse: [];
  Voltage_CurrentrTipahse: [];
}